Safari-611.1.10/LayoutTests/webaudio/audiocontext-restriction-audiobuffersourcenode-start-expected.txt

Basic tests for AudioNode API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS context.state is "suspended"
node.connect(context.destination)
PASS context.state is "suspended"
Calling context.resume() without a user gesture
Wait for 100ms
Calling node.start() with a user gesture
PASS context.resume() (without a user gesture) promise resolved
PASS context.state is "running"
PASS successfullyParsed is true

TEST COMPLETE
